+++ web/db/models/Account.ts
@@ -1,16 +1,6 @@
 import mongoose from "mongoose";
+import { IAccount, Role } from "../../utils/types/account";
 const { Schema } = mongoose;
-
-export interface IAccount {
-  email: string;
-  role: Role;
-}
-
-export enum Role {
-  Admin = "admin",
-  ContentCreator = "contentCreator",
-  Volunteer = "volunteer",
-}
 
 const accountSchema = new Schema<IAccount>({
   email: {
