+++ web/utils/types/account.ts
@@ -1,0 +1,10 @@
+export interface IAccount {
+  email: string;
+  role: Role;
+}
+
+export enum Role {
+  Admin = "admin",
+  ContentCreator = "contentCreator",
+  Volunteer = "volunteer",
+}
