+++ web/utils/types/user.ts
@@ -1,0 +1,9 @@
+import { Role } from "./account";
+
+export interface IUser {
+  email: string;
+  name: string;
+  uid: string;
+  role: Role;
+  disabled: boolean;
+}
